debut freeze

+++ terrainelement.py
@@ -5,6 +5,7 @@ import random
 
 
 class TerrainElement(ABC):  # abstract class
+
     def __init__(self, x, y, scale, folder, filename):
         # made with a method so that it can be set in each class
         self.folder = folder
@@ -14,7 +15,7 @@ class TerrainElement(ABC):  # abstract class
         self.toggleSprite("night")
         # self.setScale(self.scale)  # change the size of the object
         self.rect = self.image.get_rect(x=x, y=y)
-
+        self.freeze=False
     def move(self, xDirection, yDirection):
         self.rect.move_ip(xDirection, yDirection)
 
@@ -50,8 +51,9 @@ class Obstacle(TerrainElement):
             terrain.effects.append(Smoke(terrain, x - 60, y - 80))
 
     def update(self, player):
-        if self.rect.colliderect(player.rect):
-            player.bounce()
+        if self.freeze == False :
+            if self.rect.colliderect(player.rect):
+                player.bounce()
 
 
 class Fruit(TerrainElement):
@@ -66,6 +68,8 @@ class Fruit(TerrainElement):
         elif randomfruit == 1:
             filename = "apple"
         elif randomfruit == 2:
+            filename = "apple"
+        elif randomfruit == 3:
             filename = "cherry"
 
         TerrainElement.__init__(self, x, y, scale, folder, filename)
@@ -81,6 +85,9 @@ class Fruit(TerrainElement):
                 scoreAmount = 4
             elif self.randomfruit == 2:
                 player.recoveryhp()
+                scoreAmount = 2
+            elif self.randomfruit == 3:
+                player.addFreezeall()
                 scoreAmount = 2
             player.addScore(scoreAmount)
             self.getEaten()
@@ -150,13 +157,18 @@ class Projectile(TerrainElement):
         TerrainElement.__init__(self, x, y, scale, folder, filename)
 
     def update(self, player):
-        self.rect.move_ip(self.velocity, 0)
-        if self.rect.colliderect(player.rect):
-            player.loseHP()
-            # projectile deletes itself when it touches character
-            self.delete()
-        if pygame.time.get_ticks() > self.creationTime + self.expirationDate:
-            self.delete()
-
+        if self.freeze == False :
+            self.rect.move_ip(self.velocity, 0)
+            if self.rect.colliderect(player.rect):
+                player.loseHP()
+                # projectile deletes itself when it touches character
+                self.delete()
+            if pygame.time.get_ticks() > self.creationTime + self.expirationDate:
+                self.delete()
+        else :
+            currentTime = pygame.time.get_ticks()
+            print(pygame.time.get_ticks())
+            if currentTime > 15000 :
+                self.freeze = False
     def delete(self):
         self.game.terrain.projectiles.remove(self)
